More developed version for WinBoard Alien Edition
This commit contains the code as it was further developed on Windows,
and tested under WB-Alien. Highlighting was added, and a 3-ply fixed-depth
search is invoked when it is the engines turn to move. Piece-square tables
have been added to draw pieces to the center, and the move generator respects
the special Lion-capture and promotion rules of Chu Shogi. Most of the time
things work, but there must be a bad bug, as sometimes it suddenly moves
with an opponent piece!