Give pieces of same type same Zobrist key
Give different piece keys to white and black pieces of same type. Switch
to a system that assigns the key 'on demand', when the entry in the
list of piece descriptions is first used. Let Lookup return pointer
rather than number in list, so that future versions can search in multiple
lists (for other variants).