Add null move
The WB Alien null-move notation is recognized on input, and generated
on output. The internal representation is a move towards the same square
by a piece that has this capability. (This currently fails on a1, due
to 0 being used as representation for invalid moves!)
Only one null move is generated and added to the move list.
The null move is not searched yet.