Implement Dai Shogi
The board format was made variable, and the relevant parameters are set
during Init() according to the current variant (now passed as parameter).
LookUp() now searches the requested piece in multiple lists, depending
on the current variant. E.g. for Dai it searches daiPieces, followed by
chuPieces. The WB variant command is now interpreted, by comparing
the mentioned variat to the list of variants. Support for plain Knight
moves was added. The suppression of promotions after an initial deferral
was made subject to chuFlag.