Use setup command in Sho Shogi
To exploit the new engine-defined-variants mechanism of XBoard,
a duplicate of Sho Shogi is added with an engine-defined name.
The length of the variant table is now determined dynamically,
by using the sentinel, and the variants feature is now generated from
this table (rather than hard-coded). The variant 12x12+0_fairy has
been dropped as alternative for Chu Shogi, as this never really worked,
and chu has been elevated to standard variant in the XBoard main branch.