Fix royalty of Crown Prince in setup positions
The Crown Prince wa not recognized as royal, because the tet supposed to
do that assumed it was CP, rather than +DE. During initialization of games
without Elephant this caused insertion of a dummy (captured) +DE as second
royal, demoting a possibly initially present +DE to non-royal status.
 Testing exposed an bug in the evaluation that wrongly identified CP
in the abense of K, leading to a segfault.