Let heavy steppers form fortress
The heavy steppers now use the old stepper table with pre-promotion bonus,
but made a little bit more neutral in the area of the fortress, to prevent
them from early wandering off. A light fortress bonus, which is slowly
switched on during the game, and dependent on opponent Lion presence,
ties them down into a fortress.