Fix hash key after setboard
The initial hash key was the same in any start position. This resulted
in a has-hit on the opening position, which was searched to make a move list
after 'new', in the position after 'setboard', so the opening's hash move
could be played in any position, even where it was completely invalid.