Fix double-capture of royal pair
Kings could step in check if there was a Crown Prince, but there was
no allowance for the fact that this could expose both of them to
simultaneous capture by a Lion. So the search can visit nodes where there
is no royal, which did cause a cratch. This is now intercepted at the top
of Search, in the check test, which aborts the branch if there is no royal.
Except in tsume for the winning side. (Although this then might still
cause the crash, probably in eval.)