Fix phantom move when no non-captures
When there were no non-captures generated at all, a random one would be
tried that happened to be left on the move stack from another node, which
needed not even be a faulty move. In some tsume problems where this move
happened to be the capture of the King with an empty square, this lead to
reporting false mates. (Usually you only have no non-captures if you have
no pieces, and in tsume you might not even have a King.)