Fix setting up of positions without King
Without King, royal[] would not be set, and retain the value it had
in the standard setup (as it was not cleared either). This would usually
point outside the piece list, but occasionally it designated an on-board
piece as royal, which then could make HaChu think it was checkmated in
the initial position of a tsume problem. It also caused problems
in positions with only CP.