Fix tsume condition when both K and CP present
A position where the side-to-be-mated has both K and CP is never
a check, and should thus not be allowed to occur in a tsume problem.
This was not properly detected. (Strictly speaking it can be a check
if both can be captured on the same move by a Lion/Falcon/Eagle,
but for now we ignore that.)