Use returned depth in hash table
The depth returned by the search is now also hashed. This gives an
enormous speedup of tsume searches, as mates are assigned 'infinite' depth,
and thus forced distant mates also, so that they remain valid in
subsequent iterations without re-search. For safety the depth assigned to
repetitions was limited to the current iteration.