Allow user castling in Mighty Lion
A large part of the castling code is implemented: MakeMove can perform it,
ParseMove can recognize it. The routine GenCastlings generates them,
but without paying attention to (moving through) check, and Highlight
marks the King's to-square in the color FEN.
 The castlings are only generated by ListMoves, added to the list used
for legality testing of input moves, though (and then only in variant lion).
Not yet in GenNonCaptures, so HaChu would never castle itself, but will
accept it if the user does so. It would even accept castling through check,
but the idea is that legality checking in the GUI would prevent the user
from entering that.