Implement Werewolf Chess
The Werewolf is marked as contageous by setting ranking 5, immunity
(= royalty) by ranking 4. Although promotions to Werewolf are implied,
they are explicitly written on moves, and to-squares of such moves
are highlighted in purple, to allow playing this as engine-defined
variant, where the GUI would not be able to make the implication.