Fix some compiler warnings
The -Wall option is added to the CFLAGS, and the resulting warnings
about unused variables are fixed. -Wno-parentheses is still needed to
suppress the pedantic remarks about parentheses (and was added in the
Makefile too). Only 4 complaints about possibly unitialized variables
in Search() remain; these are all false, but since this is a time-
critical routine, adding dummy assignments to suppress them is not
considered worth it.